Add files via upload
The SolarWatch ML Model is an advanced machine learning-based system designed to detect and diagnose potential issues with the performance of solar panels. It uses a combination of  data analysis, and AI algorithms to monitor and analyze the output of each solar panel in real-time whether it is faulty or Not

The SolarWatch ML Model works by using historical data to train a machine learning model to recognize patterns in the data that indicate a particular type of failure. 
Once the SolarWatch ML Model detects an issue, it sends an alert to the system administrator, who can then take appropriate action to resolve the issue. Additionally, the model continues to learn and adapt to new data, further improving its accuracy and ability to detect potential issues.
Overall, the SolarWatch ML Model is a cutting-edge tool for anyone looking to optimize the performance of their solar panel system and ensure it continues to operate at maximum efficiency while minimizing downtime and maintenance costs.